refactor: Repository interface 추가 및 기존 repository를 구현체로 변경
