+++ src/main/java/roomescape/reservation/repository/H2ReservationRepository.java
@@ -1,0 +1,66 @@
+package roomescape.reservation.repository;
+
+import java.time.LocalDate;
+import java.time.LocalTime;
+import java.util.HashMap;
+import java.util.List;
+import java.util.Map;
+import javax.sql.DataSource;
+import org.springframework.jdbc.core.JdbcTemplate;
+import org.springframework.jdbc.core.simple.SimpleJdbcInsert;
+import org.springframework.stereotype.Repository;
+import roomescape.reservation.model.Reservation;
+import roomescape.reservation.model.ReservationTime;
+
+@Repository
+public class H2ReservationRepository implements ReservationRepository{
+
+    private final JdbcTemplate jdbcTemplate;
+    private final SimpleJdbcInsert simpleJdbcInsert;
+
+    public H2ReservationRepository(DataSource dataSource) {
+        this.jdbcTemplate = new JdbcTemplate(dataSource);
+        this.simpleJdbcInsert = new SimpleJdbcInsert(jdbcTemplate)
+                .withTableName("reservation")
+                .usingGeneratedKeyColumns("id");
+    }
+
+    @Override
+    public List<Reservation> findAll() {
+        String sql = """
+                SELECT
+                    r.id as reservation_id,
+                    r.name,
+                    r.date,
+                    t.id as time_id,
+                    t.start_at as time_value
+                FROM reservation as r
+                inner join reservation_time as t
+                on r.time_id = t.id
+                """;
+        return jdbcTemplate.query(sql, (resultSet, rowNum) -> new Reservation(
+                resultSet.getLong("reservation_id"),
+                resultSet.getString("name"),
+                resultSet.getObject("date", LocalDate.class),
+                new ReservationTime(resultSet.getLong("time_id"),
+                        resultSet.getObject("start_at", LocalTime.class))
+        ));
+    }
+
+    @Override
+    public Reservation insertReservation(Reservation reservation) {
+        Map<String, Object> parameters = new HashMap<>();
+        parameters.put("name", reservation.getName());
+        parameters.put("date", reservation.getDate());
+        parameters.put("time_id", reservation.getTime().getId());
+        Number number = simpleJdbcInsert.executeAndReturnKey(parameters);
+        return new Reservation(number.longValue(), reservation.getName(), reservation.getDate(), reservation.getTime());
+    }
+
+    @Override
+    public boolean deleteReservationById(long id) {
+        String sql = "delete from reservation where id = ?";
+        int updated = jdbcTemplate.update(sql, id);
+        return updated != 0;
+    }
+}
