+++ src/main/java/roomescape/reservation/repository/H2ReservationTimeRepository.java
@@ -1,0 +1,59 @@
+package roomescape.reservation.repository;
+
+import java.time.LocalTime;
+import java.util.HashMap;
+import java.util.List;
+import java.util.Map;
+import java.util.Optional;
+import javax.sql.DataSource;
+import org.springframework.jdbc.core.JdbcTemplate;
+import org.springframework.jdbc.core.simple.SimpleJdbcInsert;
+import org.springframework.stereotype.Repository;
+import roomescape.reservation.model.ReservationTime;
+
+@Repository
+public class H2ReservationTimeRepository implements ReservationTimeRepository{
+
+    private final JdbcTemplate jdbcTemplate;
+    private final SimpleJdbcInsert simpleJdbcInsert;
+
+    public H2ReservationTimeRepository(DataSource dataSource) {
+        this.jdbcTemplate = new JdbcTemplate(dataSource);
+        this.simpleJdbcInsert = new SimpleJdbcInsert(dataSource)
+                .withTableName("reservation_time")
+                .usingGeneratedKeyColumns("id");
+    }
+
+    @Override
+    public ReservationTime insertTime(ReservationTime reservationTime) {
+        Map<String, Object> parameters = new HashMap<>();
+        parameters.put("start_at", reservationTime.getStartAt());
+        Number number = simpleJdbcInsert.executeAndReturnKey(parameters);
+        return new ReservationTime(number.longValue(), reservationTime.getStartAt());
+    }
+
+    @Override
+    public List<ReservationTime> findAll() {
+        String sql = "Select * from reservation_time";
+        return jdbcTemplate.query(sql, (resultSet, rowNum) -> new ReservationTime(
+                resultSet.getLong("id"),
+                resultSet.getObject("start_at", LocalTime.class)
+        ));
+    }
+
+    @Override
+    public boolean deleteTimeById(long id) {
+        String sql = "delete from reservation_time where id = ?";
+        int updated = jdbcTemplate.update(sql, id);
+        return updated != 0;
+    }
+
+    @Override
+    public Optional<ReservationTime> findById(long id) {
+        String sql = "select * from reservation_time where id = ?";
+        return jdbcTemplate.query(sql, (resultSet, rowNum) -> new ReservationTime(
+                resultSet.getLong("id"),
+                resultSet.getObject("start_at", LocalTime.class)
+        ), id).stream().findFirst();
+    }
+}
